requirements.txt
Flask
flask-cors

gunicorn
Flask-SocketIO
requests
requests-ntlm
uvicorn

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client